requirements.txt
Flask
boto3
pillow
pyzbar
Pillow
python-dotenv
flask_cors